+++ Facturacion/DTOs/UpdateSellerDto.cs
@@ -4,7 +4,7 @@
   {
     public string? FirstName { get; set; }
     public string? LastName { get; set; }
-    public byte? CommissionPercentage { get; set; }
+    public int? CommissionPercentage { get; set; }
     public bool? IsActive { get; set; }
   }
 }
