+++ Facturacion/Models/Seller.cs
@@ -5,7 +5,7 @@
     public int Id { get; set; }
     public required string FirstName { get; set; }
     public required string LastName { get; set; }
-    public byte CommissionPercentage { get; set; }
+    public int CommissionPercentage { get; set; }
     public bool IsActive { get; set; }
   }
 }
