+++ Facturacion/Validators/UpdateSellerValidator.cs
@@ -16,7 +16,7 @@ namespace Facturacion.Validators
         .WithMessage("El apellido es obligatorio");
 
       RuleFor(s => s.CommissionPercentage)
-        .InclusiveBetween((byte)0, (byte)100)
+        .InclusiveBetween(0, 100)
         .WithMessage("El porcentaje de comisión debe estar entre 0 y 100")
         .When(s => s.CommissionPercentage.HasValue);
